api/dev_requirements.txt
# Only development requirements
python-dotenv